Added support for Alternate Stylesheets
Alternative Stylesheets may be offered by a website to allow users to customize the browsing experience or to enhance accessibility. They are susceptible to the same attack vectors as preferred or persistent stylesheets and should therefore also make use of SRI hashes.

This change merely enables the regular expression which identifies stylesheet LINK tags to recognize alternative (alternate) stylesheets.